Moved ServiceRequestData to new Entities package
