+++ src/main/java/edu/wpi/teame/controllers/IRequestController.java
@@ -1,5 +1,7 @@
 package edu.wpi.teame.controllers;
 
+import edu.wpi.teame.entities.ServiceRequestData;
+
 public interface IRequestController {
 
   ServiceRequestData sendRequest();
